refactor richtext export and import
