+++ components/RichText/index.ts
@@ -1,0 +1,3 @@
+import RichText from './RichText';
+
+export default RichText;
