+++ pages/jobs/[slug].tsx
@@ -8,7 +8,7 @@ import { getJobPage } from 'lib/api';
 import Layout from 'components/Layout';
 import ClientLogos from 'components/ClientLogos';
 import Benefits from 'components/Benefits';
-import RichText from 'components/RichText/RichText';
+import RichText from 'components/RichText';
 import { ApplyButton } from 'components/Button';
 
 type JobPageProps = {
